+++ doc/src/release_notes.rst
@@ -46,9 +46,12 @@ To see all issues & pull requests closed by this release see the `Git closed mil
 
 * `#2057 <https://github.com/pgRouting/pgrouting/issues/2057>`__: trspViaEdges columns in different order
 * `#2087 <https://github.com/pgRouting/pgrouting/issues/2087>`__: pgr_extractVertices to proposed
+* `#2203 <https://github.com/pgRouting/pgrouting/issues/2087>`__: pgr_dijkstraNear and pgr_dijkstraNearCost to proposed
 
-.. rubric:: Experimental moved to proposed
+.. rubric:: Experimental promoted to proposed
 
+* pgr_dijkstraNear
+* pgr_dijkstraNearCost
 * pgr_extractVertices
 
 pgRouting 3.2.2 Release Notes